Correct: Function to find sibling node
Previously, the function binary_tree_sibling() did not check if the
parent node had two child nodes. Moreover, it did not also take into
consideration whether the node whose sibling is being found, is the
parent's left or right child.
The changes made ensure that the parent node has two children, before
checking if the subject node is on the right or left subtree of its
parent.