Correct: Insert a right or left child node
Previously, the function logic resulted in a memory leak if the parent
node was a NULL reference. This is because memory for the new node to be
added was being allocated before a NULL-check was performed on the
parent reference.
The changes made ensure that memory if freed before the function is
cleared from the call stack.